add DavitPath.css and moved DavitPath in atoms
