+++ src/components/atomic/svg/DavitPath.css
@@ -1,4 +1,3 @@
-
 .carvPath {
     stroke-width: 2px;
     fill: transparent;
